Conflicts:
	ankiqt/ui/main.py

+++ mac/make.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 cd /Volumes/Two/anki
-echo "cleaning up..."
-rm -rf build dist
+if [ "x$debug" = "x" ]; then
+    echo "cleaning up..."
+    rm -rf build dist
+fi
 echo "syncing updates..."
 rsync -av reflex:Lib/code/libanki --exclude .git --exclude build --exclude dist --delete .
 rsync -av reflex:Lib/code/ankiqt --exclude .git --exclude build --exclude dist  --delete .
@@ -16,7 +18,9 @@ cp -Rvf /usr/local/share/kakasi/itaijidict kakasi
 # mkdir -p audio
 # cp -Rvf /usr/local/bin/lamex audio
 echo "adding image formats..."
-cp -Rvf imageformats ankiqt
+rm -rf ankiqt/imageformats
+mkdir ankiqt/imageformats
+cp -Rvf imageformats/libq{gif,jpeg,svg,tiff}* ankiqt/imageformats
 echo "building..."
 PYTHONPATH=ankiqt:libanki python ankiqt/mac/setup.py bdist_dmg
